finish implementing Builder Pattern

+++ spring-boot-project/src/main/java/com/javaweb/repository/entity/BuildingEntity.java
@@ -1,17 +1,17 @@
 package com.javaweb.repository.entity;
 
 public class BuildingEntity {
-	private long id;
+	private Long id;
 	private String name;
 	private String street;
 	private String ward;
-	private long districtId;
+	private Long districtId;
 	private String structure;
-	private int numberOfBasement;
-	private int floorArea;
+	private Long numberOfBasement;
+	private Long floorArea;
 	private String direction;
 	private String level;
-	private int rentPrice;
+	private Long rentPrice;
 	private String rentPriceDescription;
 	private String serviceFee;
 	private String carFee;
@@ -35,11 +35,11 @@ public class BuildingEntity {
 	private String managerName;
 	private String managerPhoneNumber;
 
-	public long getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -67,11 +67,11 @@ public class BuildingEntity {
 		this.ward = ward;
 	}
 
-	public long getDistrictId() {
+	public Long getDistrictId() {
 		return districtId;
 	}
 
-	public void setDistrictId(long districtId) {
+	public void setDistrictId(Long districtId) {
 		this.districtId = districtId;
 	}
 
@@ -83,19 +83,19 @@ public class BuildingEntity {
 		this.structure = structure;
 	}
 
-	public int getNumberOfBasement() {
+	public Long getNumberOfBasement() {
 		return numberOfBasement;
 	}
 
-	public void setNumberOfBasement(int numberOfBasement) {
+	public void setNumberOfBasement(Long numberOfBasement) {
 		this.numberOfBasement = numberOfBasement;
 	}
 
-	public int getFloorArea() {
+	public Long getFloorArea() {
 		return floorArea;
 	}
 
-	public void setFloorArea(int floorArea) {
+	public void setFloorArea(Long floorArea) {
 		this.floorArea = floorArea;
 	}
 
@@ -115,11 +115,11 @@ public class BuildingEntity {
 		this.level = level;
 	}
 
-	public int getRentPrice() {
+	public Long getRentPrice() {
 		return rentPrice;
 	}
 
-	public void setRentPrice(int rentPrice) {
+	public void setRentPrice(Long rentPrice) {
 		this.rentPrice = rentPrice;
 	}
